@@ -1,12 +1,12 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>cf80adf9-ce70-4ce9-bb4e-a73887a492a6</UserSecretsId>
-    <Version>1.1.115</Version>
+	  <TargetFramework>net5.0</TargetFramework>
+	  <UserSecretsId>cf80adf9-ce70-4ce9-bb4e-a73887a492a6</UserSecretsId>
+    <Version>1.1.120</Version>
     <Authors>Alva Chien</Authors>
     <Product>H.I.H.</Product>
-    <Copyright>(C) Alva Chien, 2020</Copyright>
+    <Copyright>(C) Alva Chien, 2021</Copyright>
     <PackageProjectUrl>https://github.com/alvachien/achihapi</PackageProjectUrl>
     <PackageLicenseUrl></PackageLicenseUrl>
     <Description>Web API for Home Info. Hub, built on OData and ASP.NET Core;</Description>
@@ -37,12 +37,12 @@
     <None Remove="Sqls\Delta\v9.sql" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.11" />
-    <PackageReference Include="Microsoft.AspNetCore.OData" Version="7.5.5" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.11" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.1.11" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.4" />
-    <PackageReference Include="Serilog.AspNetCore" Version="3.4.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.5" />
+    <PackageReference Include="Microsoft.AspNetCore.OData" Version="8.0.0-rc" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.5" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="5.0.0" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.2" />
+    <PackageReference Include="Serilog.AspNetCore" Version="4.1.0" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Sqls\Delta\v1.sql" />
