@@ -1,9 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+	  <TargetFramework>net5.0</TargetFramework>
 
-    <IsPackable>false</IsPackable>
+	  <IsPackable>false</IsPackable>
 
     <Version>1.1.88</Version>
 
