@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.IO;
 using System.Collections.Generic;
-using Microsoft.AspNet.OData;
+using Microsoft.AspNetCore.OData.Routing.Controllers;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
@@ -13,6 +13,9 @@ using hihapi.Utilities;
 using Microsoft.AspNetCore.Authorization;
 using System;
 using hihapi.Exceptions;
+using Microsoft.AspNetCore.OData.Query;
+using Microsoft.AspNetCore.OData.Formatter;
+using Microsoft.AspNetCore.OData.Deltas;
 
 namespace hihapi.Controllers
 {
