@@ -1,7 +1,6 @@
 ﻿using System.Linq;
 using System.IO;
 using System.Collections.Generic;
-using Microsoft.AspNet.OData;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
@@ -15,7 +14,10 @@ using System;
 using hihapi.Exceptions;
 using Microsoft.Extensions.Logging;
 using Serilog;
-using Microsoft.AspNet.OData.Routing;
+using Microsoft.AspNetCore.OData.Routing.Controllers;
+using Microsoft.AspNetCore.OData.Query;
+using Microsoft.AspNetCore.OData.Results;
+using Microsoft.AspNetCore.OData.Formatter;
 
 namespace hihapi.Controllers
 {
