@@ -1,28 +1,18 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.AspNet.OData.Extensions;
-using Microsoft.AspNet.OData.Builder;
-using Microsoft.AspNet.OData.Batch;
 using hihapi.Models;
-using Microsoft.AspNetCore.Routing;
 using System.IO;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.FileProviders;
-using Serilog;
 using Microsoft.IdentityModel.Tokens;
+using Microsoft.OData.Edm;
+using Microsoft.AspNetCore.OData;
+using Microsoft.Extensions.Hosting;
 
 namespace hihapi
 {
@@ -49,7 +39,7 @@ namespace hihapi
                     Directory.CreateDirectory(BlogFolder);
                 }
             }
-            catch(Exception exp)
+            catch(Exception)
             {
                 // Do nothing
             }
@@ -64,6 +54,8 @@ namespace hihapi
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddCors();
+
 #if DEBUG
             this.ConnectionString = Configuration["hihapi:ConnectionString"];
             System.Diagnostics.Debug.WriteLine(this.ConnectionString);
@@ -81,25 +73,23 @@ namespace hihapi
                     options.UseSqlServer(this.ConnectionString));
             }
 
-            //services.AddMvc(x => x.EnableEndpointRouting = false).SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-            services.AddAuthorization();
+            services.AddHttpContextAccessor();
+
+            services.AddControllers();
+
+            IEdmModel model = EdmModelBuilder.GetEdmModel();
+
+            services.AddOData(opt => opt.Count().Filter().Expand().Select().OrderBy().SetMaxTop(50)
+                .AddModel(model)
+                .AddModel("v1", model)
+                // .AddModel("v2{data}", model2, builder => builder.AddService<ODataBatchHandler, DefaultODataBatchHandler>(Microsoft.OData.ServiceLifetime.Singleton))
+                // .ConfigureRoute(route => route.EnableQualifiedOperationCall = false) // use this to configure the built route template
+                );
+
+            //services.AddSwaggerGen();
 
             if (Environment.EnvironmentName == "IntegrationTest")
             {
-                //services.AddAuthentication(IdentityServerAuthenticationDefaults.AuthenticationScheme)
-                //    .AddIdentityServerAuthentication(IdentityServerAuthenticationDefaults.AuthenticationScheme,
-                //        jwtOptions =>
-                //        {
-                //            // jwt bearer options
-                //            jwtOptions.Authority = "http://localhost:5005";
-                //            jwtOptions.Audience = "api.hih";
-                //            jwtOptions.RequireHttpsMetadata = false;
-                //            jwtOptions.SaveToken = true;
-                //        },
-                //        referenceOptions =>
-                //        {
-                //            // oauth2 introspection options
-                //        });
                 services.AddAuthentication("Bearer")
                     .AddJwtBearer("Bearer", options =>
                     {
@@ -115,7 +105,6 @@ namespace hihapi
                 services.AddAuthentication("Bearer")
                     .AddJwtBearer("Bearer", options =>
                     {
-                        // options.Authority = "https://localhost:41016";
                         options.Authority = "https://localhost:44353";
                         options.RequireHttpsMetadata = false;
 
@@ -125,47 +114,11 @@ namespace hihapi
                         };
                     });
 
-                //// adds an authorization policy to make sure the token is for scope 'api1'
-                //services.AddAuthorization(options =>
-                //{
-                //    options.AddPolicy("ApiScope", policy =>
-                //    {
-                //        policy.RequireAuthenticatedUser();
-                //        policy.RequireClaim("scope", "api1");
-                //    });
-                //});
-
-                //services.AddAuthentication("Bearer")
-                //    .AddJwtBearer("Bearer", options =>
-                //    {
-                //        options.Authority = "http://localhost:41016";
-                //        options.RequireHttpsMetadata = false;
-
-                //        options.Audience = "api.hih";
-                //    });
-
-                //services.AddAuthentication(IdentityServerAuthenticationDefaults.AuthenticationScheme)
-                //    .AddIdentityServerAuthentication(IdentityServerAuthenticationDefaults.AuthenticationScheme,
-                //        jwtOptions =>
-                //        {
-                //            // jwt bearer options
-                //            jwtOptions.Authority = "http://localhost:41016";
-                //            jwtOptions.Audience = "api.hih";
-                //            jwtOptions.RequireHttpsMetadata = false;
-                //            jwtOptions.SaveToken = true;
-                //        },
-                //        referenceOptions =>
-                //        {
-                //            // oauth2 introspection options
-                //        });
-
                 services.AddCors(options =>
                 {
                     options.AddPolicy(MyAllowSpecificOrigins, builder =>
                     {
                         builder.WithOrigins(
-                            "http://localhost:29521",   // AC HIH
-                            "http://localhost:29525",   // acblog
                             "https://localhost:29521",  // AC HIH
                             "https://localhost:29525"   // acblog
                         )
@@ -183,7 +136,7 @@ namespace hihapi
 #if USE_ALIYUN
                         options.Authority = "https://www.alvachien.com/idserver";
 #elif USE_AZURE
-                        options.Authority = "http://localhost:41016";
+                        options.Authority = "https://acidsrv.azurewebsites.net";
 #endif
                         options.RequireHttpsMetadata = false;
                         options.SaveToken = true;
@@ -214,8 +167,7 @@ namespace hihapi
                 });
             }
 
-            services.AddOData();
-            services.AddMvc(options => options.EnableEndpointRouting = false);
+            services.AddAuthorization();
 
             // Response Caching
             services.AddResponseCaching();
@@ -226,20 +178,17 @@ namespace hihapi
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (env.EnvironmentName == "Development")
+            if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
-            else
-            {
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
 
-#if DEBUG
-#else
-            app.UseSerilogRequestLogging();
-#endif
+            app.UseCors(MyAllowSpecificOrigins);
+
+            app.UseHttpsRedirection();
+
+            app.UseODataBatching();
+
 
             // app.UseHttpsRedirection();
             app.UseAuthentication();
@@ -248,139 +197,24 @@ namespace hihapi
                 app.UseCors(MyAllowSpecificOrigins);
             }
 
-            ODataModelBuilder modelBuilder = new ODataConventionModelBuilder(app.ApplicationServices);
-            modelBuilder.EntitySet<Currency>("Currencies");
-            modelBuilder.EntitySet<Language>("Languages");
-            modelBuilder.EntitySet<DBVersion>("DBVersions");
-            modelBuilder.EntitySet<CheckVersionResult>("CheckVersionResult");
-            modelBuilder.EnumType<HomeMemberRelationType>();
-            modelBuilder.EntitySet<HomeDefine>("HomeDefines");
-            modelBuilder.EntitySet<HomeMember>("HomeMembers");
-            modelBuilder.EntitySet<LearnCategory>("LearnCategories");
-            modelBuilder.EntitySet<LearnObject>("LearnObjects");
-            modelBuilder.EntitySet<FinanceAccountCategory>("FinanceAccountCategories");
-            modelBuilder.EntitySet<FinanceAssetCategory>("FinanceAssetCategories");
-            modelBuilder.EntitySet<FinanceDocumentType>("FinanceDocumentTypes");
-            modelBuilder.EntitySet<FinanceTransactionType>("FinanceTransactionTypes");
-            modelBuilder.EnumType<FinanceAccountStatus>();
-            modelBuilder.EnumType<RepeatFrequency>();
-            modelBuilder.EnumType<LoanRepaymentMethod>();
-            modelBuilder.EnumType<FinancePlanTypeEnum>();
-            modelBuilder.EntitySet<FinanceControlCenter>("FinanceControlCenters");
-            modelBuilder.EntitySet<FinanceAccount>("FinanceAccounts");
-            modelBuilder.EntitySet<FinanceAccountExtraDP>("FinanceAccountExtraDPs");
-            modelBuilder.EntitySet<FinanceAccountExtraAS>("FinanceAccountExtraASs");
-            modelBuilder.EntitySet<FinanceTmpDPDocument>("FinanceTmpDPDocuments");
-            modelBuilder.EntitySet<FinanceTmpLoanDocument>("FinanceTmpLoanDocuments");
-            modelBuilder.EntitySet<FinanceOrder>("FinanceOrders");
-            modelBuilder.EntitySet<FinanceOrderSRule>("FinanceOrderSRules");
-            modelBuilder.EntitySet<FinancePlan>("FinancePlans");
-            modelBuilder.EntitySet<FinanceDocument>("FinanceDocuments");
-            modelBuilder.EntitySet<FinanceDocumentItem>("FinanceDocumentItems");
-            modelBuilder.EntitySet<FinanceDocumentItemView>("FinanceDocumentItemViews");
-            var acntBalRpt = modelBuilder.EntitySet<FinanceReportByAccount>("FinanceReportByAccounts");
-            acntBalRpt.EntityType.HasKey(p => new {
-                p.HomeID,
-                p.AccountID,
-            });
-            var ccBalRpt = modelBuilder.EntitySet<FinanceReportByControlCenter>("FinanceReportByControlCenters");
-            ccBalRpt.EntityType.HasKey(p => new
-            {
-                p.HomeID,
-                p.ControlCenterID,
-            });
-            var ordBalRpt = modelBuilder.EntitySet<FinanceReportByOrder>("FinanceReportByOrders");
-            ordBalRpt.EntityType.HasKey(p => new
-            {
-                p.HomeID,
-                p.OrderID,
-            });
-            modelBuilder.ComplexType<RepeatedDates>();
-            modelBuilder.ComplexType<RepeatDatesCalculationInput>();
-            modelBuilder.ComplexType<RepeatedDatesWithAmount>();
-            modelBuilder.ComplexType<RepeatDatesWithAmountCalculationInput>();
-            // Function on root
-            modelBuilder.Function("GetRepeatedDates")
-                .ReturnsCollection<RepeatedDates>();
-//                .Parameter<RepeatDatesCalculationInput>("input");
-            var funcbuilder = modelBuilder.Function("GetRepeatedDates2")
-                .ReturnsCollection<RepeatedDates>();
-            funcbuilder.Parameter<string>("StartDate");
-            funcbuilder.Parameter<string>("EndDate");
-            funcbuilder.Parameter<int>("RepeatType");
-            modelBuilder.Function("GetRepeatedDatesWithAmount")
-                .ReturnsCollection<RepeatedDatesWithAmount>();
-            //.Parameter<RepeatDatesWithAmountCalculationInput>("input");
-            modelBuilder.Function("GetRepeatedDatesWithAmountAndInterest")
-                .ReturnsCollection<RepeatedDatesWithAmountAndInterest>();
-                //.Parameter<RepeatDatesWithAmountAndInterestCalInput>("input");
-            // Actions on Documents
-            var docEntity = modelBuilder.EntityType<FinanceDocument>();
-            docEntity.Property(c => c.TranDate).AsDate();
-            docEntity.Collection
-                .Action("PostDPDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-            //.Returns<FinanceDocument>();
-            //.Parameter<int>("HomeID");
-            docEntity.Collection
-                .Action("PostLoanDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-//              .Parameter<int>("HomeID");
-            docEntity.Collection
-                .Action("PostAssetBuyDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-                //.Parameter<int>("HomeID");
-            docEntity.Collection
-                .Action("PostAssetSellDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-                //.Parameter<int>("HomeID");
-            docEntity.Collection
-                .Action("PostAssetValueChangeDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-            modelBuilder.Namespace = typeof(Currency).Namespace;
-            // Function on DP template documents
-            var tmpTpDocEntity = modelBuilder.EntityType<FinanceTmpDPDocument>();
-            var tmpTpDocPostFunc =
-                tmpTpDocEntity.Collection.Action("PostDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-            //tmpTpDocPostFunc.Parameter<int>("DocumentID");
-            //tmpTpDocPostFunc.Parameter<int>("AccountID");
-            //tmpTpDocPostFunc.Parameter<int>("HomeID");
-            // Action on Loan template documents: Repay document
-            var tmpLoanDocEntity = modelBuilder.EntityType<FinanceTmpLoanDocument>();
-            tmpLoanDocEntity.Collection
-                .Action("PostRepayDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-            tmpLoanDocEntity.Collection
-                .Action("PostPrepaymentDocument")
-                .ReturnsFromEntitySet<FinanceDocument>("FinanceDocuments");
-            modelBuilder.EntitySet<BlogFormat>("BlogFormats");
-            modelBuilder.EntitySet<BlogUserSetting>("BlogUserSettings");
-            modelBuilder.EntitySet<BlogCollection>("BlogCollections");
-            modelBuilder.EntitySet<BlogPost>("BlogPosts");
-            modelBuilder.EntitySet<BlogPostCollection>("BlogPostCollections");
-            modelBuilder.EntitySet<BlogPostTag>("BlogPostTags");
-            // Functions
-            var postentity = modelBuilder.EntityType<BlogPost>();
-            postentity.Function("Deploy")
-                    .Returns<string>()
-                    ;
-            postentity.Function("ClearDeploy")
-                    .Returns<string>();
-            var blogsetting = modelBuilder.EntityType<BlogUserSetting>();
-            blogsetting.Function("Deploy")
-                    .Returns<string>();
+            app.UseRouting();
 
-            var model = modelBuilder.GetEdmModel();
-            app.UseODataBatching();
-            
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+
             app.UseMiddleware(typeof(ErrorHandlingMiddleware));
 
-            app.UseMvc(routeBuilder => {
-                // and this line to enable OData query option, for example $filter
-                routeBuilder.Select().Expand().Filter().OrderBy().MaxTop(100).Count();
 
-                routeBuilder.MapODataServiceRoute("ODataRoute", "api", model);
+            app.UseEndpoints(endpoints =>
+            {
+                if (env.IsDevelopment())
+                {
+                    // A odata debuger route is only for debugger view of the all OData endpoint routing.
+                    // endpoints.MapGet("/$odata", ODataRouteHandler.HandleOData);
+                }
+
+                endpoints.MapControllers();
             });
 
             app.UseResponseCaching();
@@ -393,7 +227,6 @@ namespace hihapi
                 OnPrepareResponse = ctx =>
                 {
                     // Requires the following import:
-                    // using Microsoft.AspNetCore.Http;
                     ctx.Context.Response.Headers.Append("Cache-Control", $"public, max-age={cachePeriod}");
                 }
             });
